kumo-aws/kumo-project/requirements.txt
boto3~=1.28.0
pytest~=7.4.0
moto~=4.1.0
python-dotenv~=1.0.0
awscli-local~=0.20
scikit-learn>=1.3.2
numpy>=1.26.0
joblib>=1.3.2
